refactor: rename password input
