+++ src/components/CreatePasswordInput/utils/index.ts
