+++ src/components/CreatePasswordInput/utils/validate.ts
